refactor: make all services files names start with a capital case
