+++ src/lib/services/Delete-object.ts
