+++ src/lib/services/Read-property-multiple.ts
