+++ src/lib/services/index.ts
@@ -1,32 +1,32 @@
 export * as addListElement from './Add-list-element'
-export * as alarmAcknowledge from './alarm-acknowledge'
-export * as alarmSummary from './alarm-summary'
-export * as atomicReadFile from './atomic-read-file'
-export * as atomicWriteFile from './atomic-write-file'
-export * as covNotify from './cov-notify'
-export * as covNotifyUnconfirmed from './cov-notify'
-export * as createObject from './create-object'
-export * as deleteObject from './delete-object'
-export * as deviceCommunicationControl from './device-communication-control'
-export * as error from './error'
-export * as eventInformation from './event-information'
-export * as eventNotifyData from './event-notify-data'
-export * as getEnrollmentSummary from './get-enrollment-summary'
-export * as getEventInformation from './get-event-information'
-export * as iAm from './i-am'
-export * as iHave from './i-have'
-export * as lifeSafetyOperation from './life-safety-operation'
-export * as privateTransfer from './private-transfer'
-export * as readProperty from './read-property'
-export * as readPropertyMultiple from './read-property-multiple'
-export * as readRange from './read-range'
-export * as registerForeignDevice from './register-foreign-device'
-export * as reinitializeDevice from './reinitialize-device'
-export * as subscribeCov from './subscribe-cov'
-export * as subscribeProperty from './subscribe-property'
-export * as timeSync from './time-sync'
-export * as timeSyncUTC from './time-sync'
-export * as whoHas from './who-has'
-export * as whoIs from './who-is'
-export * as writeProperty from './write-property'
-export * as writePropertyMultiple from './write-property-multiple'
+export * as alarmAcknowledge from './Alarm-acknowledge'
+export * as alarmSummary from './Alarm-summary'
+export * as atomicReadFile from './Atomic-read-file'
+export * as atomicWriteFile from './Atomic-write-file'
+export * as covNotify from './Cov-notify'
+export * as covNotifyUnconfirmed from './Cov-notify'
+export * as createObject from './Create-object'
+export * as deleteObject from './Delete-object'
+export * as deviceCommunicationControl from './Device-communication-control'
+export * as error from './Error'
+export * as eventInformation from './Event-information'
+export * as eventNotifyData from './Event-notify-data'
+export * as getEnrollmentSummary from './Get-enrollment-summary'
+export * as getEventInformation from './Get-event-information'
+export * as iAm from './I-am'
+export * as iHave from './I-have'
+export * as lifeSafetyOperation from './Life-safety-operation'
+export * as privateTransfer from './Private-transfer'
+export * as readProperty from './Read-property'
+export * as readPropertyMultiple from './Read-property-multiple'
+export * as readRange from './Read-range'
+export * as registerForeignDevice from './Register-foreign-device'
+export * as reinitializeDevice from './Reinitialize-device'
+export * as subscribeCov from './Subscribe-cov'
+export * as subscribeProperty from './Subscribe-property'
+export * as timeSync from './Time-sync'
+export * as timeSyncUTC from './Time-sync'
+export * as whoHas from './Who-has'
+export * as whoIs from './Who-is'
+export * as writeProperty from './Write-property'
+export * as writePropertyMultiple from './Write-property-multiple'
